rename NameForm -> LoginForm
