--- src/features/Authentication/components/NameForm/index.ts
@@ -1,1 +1,0 @@
-export * from "./NameForm";
